RavenDB-25281 Code-review fixes: CompoundExact residual gate, int Limit, config-driven plan-cache sizes
- Gate CompoundExact to exactly-2-clause queries so a residual clause can never
  leak (compound key encodes only two fields); residual cases fall back to the
  bitmap pipeline. Adds CompoundExactMustApplyResidualClause regression test.
- Limit is int (RoaringBitmap.Count is int); drop the long widening in the IL
  limit-check and saturate take->int at the read-op boundary.
- Suppress the plain-leaf self-guard in AndInto so IsEmpty is emitted once.
- Plan-cache sizes (MaxPlansPerQuery / MaxDistinctQueries) are now
  ServerWideOrPerDatabaseOrPerIndex configuration, wired through CoraxIndexPersistence.

+++ src/Corax/Querying/Matches/CompiledQueryMatch.cs
@@ -68,7 +68,9 @@ public class CompiledQueryMatch(
 
     public LowLevelTransaction Llt;
 
-    public long Limit = long.MaxValue;
+    // Count is an int (RoaringBitmap.Count), so the limit can only meaningfully range over int — a
+    // larger ceiling is unreachable. Keep it int to avoid widening Count on every limit check.
+    public int Limit = int.MaxValue;
 
     public long[] Timings;  
     public long[] ResultCounts;
